backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
bcrypt==4.0.1
python-jose[cryptography]
pydantic
pydantic-settings
email-validator
fastapi-mail
passlib
python-multipart
aiofiles